# Use Vector128 and Vector256 in interface cast