Is this really needed?

Not sure how to measure the performance change. Repeat converts should be handled by the cast cache instead.