@mangod9 2023-04-03
@VSadov could you please CR this change? @huoyaoyuan, what is the motivation behind this change, have you been able to confirm perf improvements on supported hardware?