@VSadov 2023-04-04
Is this fundamentally different from vectorization that was tried in https://github.com/dotnet/runtime/pull/35468 ?