@mangod9 2023-05-01
@huoyaoyuan, without a benchmark showing improvement wouldn't be desirable to get this change in, since this is a fairly hot path.