@huoyaoyuan 2023-05-06
I tried to demonstrate the performance by commenting out all the `CastCache` related code in managed `CastHelper`.
The synthesized test code is:
```csharp
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

namespace BenchmarkGround
{
    public class Program
    {
        public readonly object obj = new C();

        static void Main(string[] args)
        {
            BenchmarkSwitcher
                .FromAssembly(typeof(Program).Assembly)
                .Run(args);
        }

        [Benchmark]
        public bool SuccessI1() => obj is I1;

        [Benchmark]
        public bool SuccessI4() => obj is I4;

        [Benchmark]
        public bool SuccessGeneric() => obj is I2<object>;

        [Benchmark]
        public bool SuccessCovariant() => obj is I3<string>;

        [Benchmark]
        public bool SuccessI15() => obj is I15;

        [Benchmark]
        public bool FailI16() => obj is I16;
    }

    interface I1 { }
    interface I2<T> { }
    interface I3<out T> { }
    interface I4 { }
    interface I5 { }
    interface I6 { }
    interface I7 { }
    interface I8 { }
    interface I9 { }
    interface I10<T> { }
    interface I11<out T> { }
    interface I12 { }
    interface I13 { }
    interface I14 { }
    interface I15 { }
    interface I16 { }

    class C :
        I1, I2<object>, I3<object>, I4, I5, I6, I7, I8,
        I9, I10<string>, I11<object>, I11<int>, I12, I13, I14, I15
    {
    }
}
```

##### Test results:

BenchmarkDotNet=v0.13.5, OS=Windows 11 (10.0.22621.1635/22H2/2022Update/SunValley2)
13th Gen Intel Core i9-13900K, 1 CPU, 32 logical and 24 physical cores
.NET SDK=8.0.100-preview.3.23178.7
  [Host]     : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2
  Job-AKCEAO : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2
  Job-CSAPOZ : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2


|           Method |        Job |                 Toolchain |      Mean |     Error |    StdDev | Ratio | RatioSD |
|----------------- |----------- |-------------------------- |----------:|----------:|----------:|------:|--------:|
|        SuccessI1 | Job-AKCEAO |   \PR-nocache\corerun.exe |  1.024 ns | 0.0070 ns | 0.0065 ns |  1.02 |    0.02 |
|        SuccessI1 | Job-CSAPOZ | \main-nocache\corerun.exe |  1.002 ns | 0.0227 ns | 0.0213 ns |  1.00 |    0.00 |
|                  |            |                           |           |           |           |       |         |
|        SuccessI4 | Job-AKCEAO |   \PR-nocache\corerun.exe |  1.319 ns | 0.0246 ns | 0.0230 ns |  0.98 |    0.02 |
|        SuccessI4 | Job-CSAPOZ | \main-nocache\corerun.exe |  1.350 ns | 0.0133 ns | 0.0125 ns |  1.00 |    0.00 |
|                  |            |                           |           |           |           |       |         |
|   SuccessGeneric | Job-AKCEAO |   \PR-nocache\corerun.exe |  1.318 ns | 0.0122 ns | 0.0114 ns |  0.96 |    0.01 |
|   SuccessGeneric | Job-CSAPOZ | \main-nocache\corerun.exe |  1.374 ns | 0.0069 ns | 0.0065 ns |  1.00 |    0.00 |
|                  |            |                           |           |           |           |       |         |
| SuccessCovariant | Job-AKCEAO |   \PR-nocache\corerun.exe | 65.537 ns | 0.4343 ns | 0.4062 ns |  1.03 |    0.01 |
| SuccessCovariant | Job-CSAPOZ | \main-nocache\corerun.exe | 63.713 ns | 0.4048 ns | 0.3786 ns |  1.00 |    0.00 |
|                  |            |                           |           |           |           |       |         |
|       SuccessI15 | Job-AKCEAO |   \PR-nocache\corerun.exe |  2.384 ns | 0.0383 ns | 0.0339 ns |  0.85 |    0.01 |
|       SuccessI15 | Job-CSAPOZ | \main-nocache\corerun.exe |  2.798 ns | 0.0473 ns | 0.0443 ns |  1.00 |    0.00 |
|                  |            |                           |           |           |           |       |         |
|          FailI16 | Job-AKCEAO |   \PR-nocache\corerun.exe |  2.532 ns | 0.0302 ns | 0.0282 ns |  0.78 |    0.01 |
|          FailI16 | Job-CSAPOZ | \main-nocache\corerun.exe |  3.240 ns | 0.0277 ns | 0.0246 ns |  1.00 |    0.00 |

Then I reverted the cache lookup code and got the performance:

|           Method |        Job |         Toolchain |      Mean |     Error |    StdDev | Ratio | RatioSD |
|----------------- |----------- |------------------ |----------:|----------:|----------:|------:|--------:|
|        SuccessI1 | Job-WBWNPA |   \PR\corerun.exe | 1.0099 ns | 0.0049 ns | 0.0043 ns |  1.03 |    0.01 |
|        SuccessI1 | Job-EPRMZY | \main\corerun.exe | 0.9796 ns | 0.0131 ns | 0.0123 ns |  1.00 |    0.00 |
|                  |            |                   |           |           |           |       |         |
|        SuccessI4 | Job-WBWNPA |   \PR\corerun.exe | 0.9982 ns | 0.0237 ns | 0.0222 ns |  0.97 |    0.02 |
|        SuccessI4 | Job-EPRMZY | \main\corerun.exe | 1.0337 ns | 0.0088 ns | 0.0082 ns |  1.00 |    0.00 |
|                  |            |                   |           |           |           |       |         |
|   SuccessGeneric | Job-WBWNPA |   \PR\corerun.exe | 1.0154 ns | 0.0086 ns | 0.0081 ns |  1.01 |    0.01 |
|   SuccessGeneric | Job-EPRMZY | \main\corerun.exe | 1.0083 ns | 0.0128 ns | 0.0120 ns |  1.00 |    0.00 |
|                  |            |                   |           |           |           |       |         |
| SuccessCovariant | Job-WBWNPA |   \PR\corerun.exe | 1.1898 ns | 0.0187 ns | 0.0175 ns |  1.01 |    0.02 |
| SuccessCovariant | Job-EPRMZY | \main\corerun.exe | 1.1751 ns | 0.0185 ns | 0.0173 ns |  1.00 |    0.00 |
|                  |            |                   |           |           |           |       |         |
|       SuccessI15 | Job-WBWNPA |   \PR\corerun.exe | 2.5057 ns | 0.0453 ns | 0.0423 ns |  0.99 |    0.02 |
|       SuccessI15 | Job-EPRMZY | \main\corerun.exe | 2.5206 ns | 0.0125 ns | 0.0111 ns |  1.00 |    0.00 |
|                  |            |                   |           |           |           |       |         |
|          FailI16 | Job-WBWNPA |   \PR\corerun.exe | 2.6555 ns | 0.0721 ns | 0.1057 ns |  0.95 |    0.03 |
|          FailI16 | Job-EPRMZY | \main\corerun.exe | 2.8266 ns | 0.0199 ns | 0.0186 ns |  1.00 |    0.00 |

It's a clear win for traversing a long interface list when the cast cache is disabled, but less impressive when the cache hits.