@huoyaoyuan 2023-05-06
After reading recent issues, I'm afraid the change would be unfriendly for R2R/Tier0 code. @stephentoub Do you have any idea here?