@stephentoub 2023-05-06
If we could reduce complexity by using an existing helper (eg SpanHelpers.ContainsValueType) and doing so doesn't hurt throughput, especially for smaller numbers of interfaces, I think that'd be ok. Otherwise, it doesn't look worthwhile.